Now using sets instead of list to store suggested cities.
Using sets in the AutocompleteTrie allowed us to prevent
duplication of city objects. However, we always needed to
copy the set into a list for the JSON response. Fortunately,
it is possible directly to map a set to a JSON list.

I also enforced a call to toLowerCase() when storing keys in
the trie.