Fixed bugs than prevented deployment
Added compiler target in pom.xml
Note: I had some issues loading my JAR in Elastic Beanstalk and the problem
was that it couldn't find my configuration file in the classpath.
Time is starting to be missing so I decided to include options directly
in the source file (I am not proud!) so I can have time to add more
features.